Set shiftwidth based on the GCD of the most common tab spacings.
The previous method of using the shortest existing tab spacing worked
some of the time, but isn't as robust as the GCD method (i.e. if there
was one or two lines that started with only one space, but the rest of
the lines started with four spaces, the old method would set shiftwidth
to 1).