Better conform to tabstop help guidelines
Always set softtabstop. This is more consistent with previous behavior
and from reading the softtabstop help seems to make more sense that we
set all of these values to the same thing.

In the case where we find both tabs and spaces, now we set expandtab as
described in #2:

	2. Set 'tabstop' and 'shiftwidth' to whatever you prefer and use
	   'expandtab'.  This way you will always insert spaces.  The
	   formatting will never be messed up when 'tabstop' is changed.

Alternatively, we could follow #1 and set tabstop=8 (instead of
guessing) and use noexpandtab:

	1. Always keep 'tabstop' at 8, set 'softtabstop' and 'shiftwidth' to 4
	   (or 3 or whatever you prefer) and use 'noexpandtab'.  Then Vim
	   will use a mix of tabs and spaces, but typing <Tab> and <BS> will
	   behave like a tab appears every 4 (or 3) characters.

That's what vim-sleuth does, so why copy their recipe? Guessing is
probably more useful to users.