Add option g:detectindent_never_mix to avoid mixed indent
In case of mixed indent (this happens in some of my source files with tab
indent and multiline strings containing spaces) I don't want to switch to
mixed indent which is not correctly detected in my case. I want to indent
with either spaces or tabs depending on the first which is seen on the
file.